+++ test/database_test.dart
@@ -1,5 +1,3 @@
-import 'dart:convert';
-
 import 'package:drift/drift.dart' hide isNull;
 import 'package:drift/native.dart';
 import 'package:flutter_test/flutter_test.dart';